chore: ResourceUtils moved from security-server to common-rest-api
Refs: XRDDEV-2295
